Move some demo files around
