+++ plugins/@grouparoo/demo/src/bin/grouparoo/demo/eventStream.ts
@@ -1,6 +1,6 @@
 import { log, CLI, config } from "actionhero";
 import { MockSession, getApiKey } from "../../../util/MockSession";
-import { getPurchaseCategories } from "../../../sample_data";
+import { getPurchaseCategories } from "../../../events/db";
 import { ApiKey, App } from "@grouparoo/core";
 
 const sleep = 100;
